AssetMovement/requirements.txt
anyascii==0.2.0
arabic-reshaper==2.1.1
asgiref==3.3.1
beautifulsoup4==4.8.2
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-crispy-forms==1.11.1
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.4.0
django-treebeard==4.5.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docx-mailmerge==0.5.0
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
future==0.18.2
html5lib==1.1
idna==2.10
l18n==2020.6.1
lxml==4.6.3
mysql-connector-python==8.0.23
mysqlclient==2.0.3
openpyxl==3.0.7
Pillow==8.1.2
PyJWT==2.0.1
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.5.65
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.4
wagtail==2.12.4
webencodings==0.5.1
whitenoise==5.2.0
Willow==1.4
xhtml2pdf==0.2.5
xlrd==2.0.1
XlsxWriter==1.3.9
xlwt==1.3.0
